clients commit

+++ src/Schd.Notification/Controllers/ClientController.cs
@@ -12,19 +12,20 @@ namespace Schd.Notification.Controllers
 {
     public class ClientController :BaseController
     {
-        private readonly AppDbContext _db;
+        private readonly IAppDbContext _db;
 
-        public ClientController(AppDbContext db)
+        public ClientController(IAppDbContext db)
         {
             _db = db;
         }
 
         [HttpPost]
-        [ProducesDefaultResponseType(typeof(ApiResponse))]
+        [ProducesDefaultResponseType(typeof(ApiResponse<string>))]
         public async Task<IActionResult> Add([FromBody] ClientModel clientModel)
         {
             var exist = _db.Clients.Any(x => x.Secret == clientModel.Secret);
-            if (!exist)
+
+            if (!exist && Guid.TryParse(clientModel.Secret, out var secret))
             {
                 var client = new Client
                 {
@@ -53,18 +54,25 @@ namespace Schd.Notification.Controllers
             return Ok(clients);
         }
 
+        [HttpGet("{id}")]
+        [ProducesDefaultResponseType(typeof(ApiResponse))]
+        public async Task<IActionResult> Approved(string id)
+        {
+            var clients = _db.Clients.FirstOrDefault(x=>$"{x.Id}" == id);
+
+            return Ok(clients);
+        }
+
         [HttpDelete]
         [ProducesDefaultResponseType(typeof(ApiResponse))]
         public async Task<IActionResult> Delete(string id)
         {
             var client = await _db.Clients.FirstOrDefaultAsync(x => x.Id.ToString() == id);
-            if (client!=null)
+            if (client!=null && Guid.TryParse(id, out var secret))
             {
                 _db.Clients.Remove(client);
                 await _db.SaveChangesAsync();
-
-
-
+                
                 return Ok("Client deleted");
             }
